Crate dnstap

An implementation of the dnstap protocol

This crate implements the sender part of the dnstap protocol, a flexible, structured binary log format for DNS software.

Structs

AuthQuery

AuthResponse

ClientQuery

ClientResponse

DNSMessage
A DNS message.

DNSTapBuilder
Builds a DNSTapWriter service.

DNSTapPendingWriter

DNSTapWriter
DNSTapWriter is responsible for receiving DNS messages, connecting (and automatically reconnecting) to a UNIX socket, and asynchronously pushing the serialized data using frame stream protocol.

ResolverQuery

ResolverResponse

Sender
Sender is a cloneable structure to send DNS messages.

Enums

MessageType

SocketFamily

SocketProtocol