Struct dnstap::DNSTapPendingWriter

pub struct DNSTapPendingWriter { /* fields omitted */ }

Methods

impl DNSTapPendingWriter

fn listen(builder: DNSTapBuilder) -> Result<DNSTapPendingWriter, &'static str>

Creates a DNSTapPendingWriter object. The communication channel is established at this point, and the sender() function can be used in order to get Sender objects.

fn start(self) -> Result<DNSTapWriter>

Spawns a new task handling writes to the socket.

fn sender(&self) -> Sender

Returns a cloneable Sender object that can used to send DNS messages.